Stories are deleted from dpmt

+++ Backend/Application/Logic/StoryLogic.cs
@@ -8,13 +8,15 @@ namespace Application.Logic;
 public class StoryLogic : IStoryLogic
 {
     private readonly IStoryDAO storyDAO;
+    private readonly IDepartmentLogic departmentLogic;
 
-    public StoryLogic(IStoryDAO storyDAO)
+    public StoryLogic(IStoryDAO storyDAO, IDepartmentLogic departmentLogic)
     {
         this.storyDAO = storyDAO;
+        this.departmentLogic = departmentLogic;
     }
     
-    public async Task<Story> CreateStoryAsync(string title, string body)
+    public async Task<Story> CreateStoryAsync(string title, string body, int departmentId)
     {
         IEnumerable<Story> existing = await GetStoriesAsync(title, body);
         if(existing != null)
@@ -22,15 +24,23 @@ public class StoryLogic : IStoryLogic
             throw new Exception($"A story with the same title and body already exists");
         }
 
-        StoryDTO story = new StoryDTO(title, body);
+        StoryDTO story = new StoryDTO(title, body, departmentId);
 
         Story created = await storyDAO.CreateStoryAsync(story);
+        await departmentLogic.AddStoryAsync(departmentId,created.id);
         return created;
     }
 
     public async Task<Story> DeleteStoryAsync(int storyId)
     {
-        return await storyDAO.DeleteStoryAsync(storyId);
+        Story? existing = await GetStoryByIdAsync(storyId);
+        if(existing == null)
+        {
+            throw new Exception($"A story with id {storyId} doesn't exist");
+        }
+        await storyDAO.DeleteStoryAsync(storyId);
+        await departmentLogic.RemoveStoryAsync(storyId);
+        return existing;
     }
 
     public async Task<IEnumerable<Story>> GetStoriesAsync(string? title, string? body)
